pyproject.toml
[project]
name = "v2ex"
version = "0.1.0"
description = "V2EX CLI client"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28",
    "typer>=0.15",
    "rich>=13",
    "python-dotenv>=1",
]

[project.scripts]
v2ex = "v2ex.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py313"
line-length = 100
